Fix for decoding pc-relative offsets and branches.
Previously the offset was assumed to be relative to 2 bytes after the
opcode, but this is incorrect. It shows itself in movem xx(pc),d0-a6
instructions where the data is

<opcode16> <reg_mask> <EA>

here the EA is 4 bytes from the opcode.
Now we always encode displacements as being relative to the instruction
opcode, and store as 32 bits (since it can overflow 16 bits in size).