Add script to generate conversion table for pmove-supporting opcodes.
We generate the test data by assembling a file of only pmove-using instructions.
Since they produce the same "top" bits (in the pmove descriptions), we can
find them in the generated .p56 file, and match them to the original source.

From this we can geneerate a table by splitting the source code opcode/operands
and translating to hopper56 enums.